move TxMeta context files to contexts/action
