+++ src/renderer/screens/Action/SignOverlay.tsx
@@ -7,7 +7,7 @@ import {
 } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { useOverlay } from '@app/contexts/Overlay';
-import { useTxMeta } from '@app/contexts/TxMeta';
+import { useTxMeta } from '@/renderer/contexts/action/TxMeta';
 import { useMemo, useRef, useState } from 'react';
 import { QRViewerWrapper } from './Wrappers';
 import { QrDisplayPayload } from '@app/library/QRCode/DisplayPayload';
